pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portfolio-agent"
version = "0.1.0"
description = "Portfolio analysis agent with codegen, sandbox execution, and caching."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml",
  "langchain>=0.2.10",
  "langchain-core>=0.2.10",
  "langchain-community>=0.2.10",
  "langchain-openai>=0.2.0",
  "langchain-google-genai>=2.0.0",
  "deepagents",
  "python-dotenv",
  "jinja2",
  "e2b-code-interpreter",
  "pandas",
  "pydantic",
  "pydantic-settings",
  "yfinance",
  "pinecone",
  "robin_stocks",
  "tavily-python",
]

[project.scripts]
portfolio-agent = "portfolio_agent.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"portfolio_agent.prompts" = ["*.j2"]
"portfolio_agent.config" = ["*.yaml"]
